fix bug 8170 to support old env var
